+++ src/babylon/logging/logger.h
@@ -24,10 +24,9 @@ class Logger final {
  private:
   using ThreadLocalLogStream =
       EnumerableThreadLocal<::std::unique_ptr<LogStream>>;
-  using Storage = ::std::array<Logger::ThreadLocalLogStream,
-                               static_cast<int>(LogSeverity::NUM)>;
-  using PointerStorage = ::std::array<::std::atomic<ThreadLocalLogStream*>,
-                                      static_cast<int>(LogSeverity::NUM)>;
+  using Storage = ::std::array<Logger::ThreadLocalLogStream, LogSeverity::NUM>;
+  using PointerStorage =
+      ::std::array<::std::atomic<ThreadLocalLogStream*>, LogSeverity::NUM>;
 
   Logger(const Logger& other) noexcept;
   Logger& operator=(const Logger& other) noexcept;
@@ -62,7 +61,7 @@ class LoggerBuilder final {
  private:
   using Storage =
       ::std::array<::std::pair<LogSeverity, Logger::ThreadLocalLogStream>,
-                   static_cast<int>(LogSeverity::NUM)>;
+                   LogSeverity::NUM>;
   Storage _log_streams;
   LogSeverity _min_severity;
 };
